attempt at invoke dynamic.. failure... try again later.

+++ src/test/java/org/boon/tests/model/SalesEmployee.java
@@ -12,7 +12,7 @@ public class SalesEmployee extends Employee implements Comparable<SalesEmployee>
     }
 
 
-    public float getCommissionRate() {
+    public int getCommissionRate() {
         return commissionRate;
     }
